requirements.txt
flask==3.1.0
flask-socketio==5.4.1
eventlet==0.37.0
python-nmap==0.7.1
dnspython==2.7.0
